drafting our training loop, continued ...
This was choking on the backwards pass because we were computing an
alleged gradient for the MLP biases that doesn't match the shape of
the biases. The solution is probably to edit MultiLayerPerceptron to
be able to take a matrix input on purpose, rather than accidentally
appearing to work due to ndarray auto-broadcasting.