Adam
Swapping this in and eyeballing our samples, this looks like it's
performing worse than our SGD implementation. It'll need tests to
verify correctness and benchmarks to check performancy. (It's possible
this CPU hobby project is in a regime where the cost of the extra
moment bookkeeping outweighs the smarter optimizer?)