remediate recursive blowup
I am feeling the AGI. I might need to quickly blog the story of this
bug; it's too illustrative of the near-future of work. I have faith
that I could have found and fixed this without LLM assistance, but I'm
pretty sure it would have taken longer.

This makes training nine times faster (with the current hyperparams,
but we might also be able to stack more layers now that they're not
exponentially expensive).