you can't build LayerNorm with those, dummy
I worked really hard yesterday on separate `SubtractMean` and
`DivideByStandardDeviation` operations, thinking that I would compose
them to build LayerNorm—and I didn't even manage to get the backward
pass of the second one correct. But it gets worse—this whole approach
is just wrong! Those operations don't compose! (The output of "x − μ"
is not the correct input to "var(x) + ε", which needs the original x.)
This isn't a subtle issue that takes years of experience to notice—a
non-retarded teenager or unusually bright child could have seen
it. I'm retarded. I'm going to die. What we need is an atomic
`Normalize` operation.