quiet compiler warnings
There were some "hiding a lifetime" warnings that I think come from
the newer compiler (1.91) than when I was last intensively working on
this in August 2024. And, I don't want to delete these structs yet.

Claude Code also did this, but I'm going to stop being neurotic about
that; I am acclimating to the world we live in now.