fix bug in tokenizer learning
I was disturbed that Claude Code spotted the bug in the previous
commit without even needing to run the test. (I had inferred that
there must be a bug due to the test assertion showing implausible
merge rules, but hadn't investigated further.) But actually, it's a
pretty dumb bug; spotting that isn't that scary.

Then I noticed that I couldn't have a stable test assertion because
the HashMap was randomizing, and Claude suggested BTreeMap. That's
also not a scary superhuman capability, even if I didn't happen to
remember BTreeMap off the top of my head.